[refactor]: Add up header components
