+++ src/constants/style.ts
@@ -1,5 +1,5 @@
 export const HEADER_ICON_BUTTON =
-  'bg-secondary-light/50 p-2 border-1 shadow-lg shadow-bg-primary/25 hover:bg-secondary-light';
+  'bg-secondary-light/50 p-2 border-1 shadow-lg shadow-bg-primary/25';
 
 export const PRIMARY_BUTTON =
   'bg-primary h-12 w-full rounded-lg text-white font-bold text-xl border-1 border-gray-light hover:bg-primary/75 shadow-lg shadow-bg-primary/25 ';
